requirements.txt
numpy>=1.24
pandas>=2.0
flask>=2.0
python-dotenv>=0.19
streamlit>=1.28
